Extract Processor as a separate class
