+++ src/main/java/com/rubensminoru/partitioners/Partitioner.java
@@ -1,0 +1,8 @@
+package com.rubensminoru.partitioners;
+
+import com.rubensminoru.messages.KafkaMessage;
+import com.rubensminoru.writers.Writer;
+
+public interface Partitioner {
+    boolean check(KafkaMessage message, Writer writer);
+}
