+++ src/main/java/com/rubensminoru/partitioners/PartitionerFactory.java
@@ -1,7 +1,7 @@
 package com.rubensminoru.partitioners;
 
 public class PartitionerFactory {
-    public TimeBasedPartitioner createInstance(String topic) {
-        return new TimeBasedPartitioner(topic);
+    public Partitioner createInstance() {
+        return new TimeBasedPartitioner();
     }
 }
