+++ src/main/java/com/rubensminoru/writers/Writer.java
@@ -1,0 +1,9 @@
+package com.rubensminoru.writers;
+
+import com.rubensminoru.messages.KafkaMessage;
+
+public interface Writer {
+    void write(KafkaMessage message);
+
+    long getTimestamp();
+}
