+++ src/main/java/com/rubensminoru/writers/WriterFactory.java
@@ -1,0 +1,7 @@
+package com.rubensminoru.writers;
+
+public class WriterFactory {
+    public Writer createInstance() {
+        return new ParquetWriter();
+    }
+}
